Rename two files to match the new enforced standard
A PR was merged to trunk enforcing the use of the kebab case pattern for
file and directory names. This commit changes the name of a directory
and two files that were created before the PR was merged.

[MAILPOET-5398]
