+++ mailpoet/assets/js/src/common/fields-section/fields-section.tsx
