+++ mailpoet/assets/js/src/segments/dynamic/form.tsx
@@ -7,8 +7,8 @@ import { Hooks } from 'wp-js-hooks';
 import { Input } from 'common/form/input/input';
 import { ReactSelect } from 'common/form/react-select/react-select';
 import { APIErrorsNotice } from 'notices/api-errors-notice';
-import { FieldsSection } from 'common/fields_section/fields_section';
-import { FieldWrapper } from 'common/fields_section/field_wrapper';
+import { FieldsSection } from 'common/fields-section/fields-section';
+import { FieldWrapper } from 'common/fields-section/field-wrapper';
 import { MailPoet } from 'mailpoet';
 import { LockedBadge } from 'common/premium-modal/locked-badge';
 import { PremiumModal } from 'common/premium-modal';
